Don't require a valid TeamProgression in TeamLoadEvent
This fixes a rare "IllegalArgumentException: Invalid TeamProgression" error when loading a player and then immediately moving them to another team. The error is caused by sync events, which need to be called on the main thread (1 tick later). In the meantime, the team provided to TeamLoadEvent might have exited the cache, like in the case described above.

For more info see: https://discord.com/channels/876458743934648340/1421261537297039381/1425497340977877147